requirements.txt
gradio>=4.16.0
ollama>=0.1.6
requests>=2.31.0
python-dotenv>=1.0.1
